refactor sso button into component
