+++ src/App.svelte
@@ -2,10 +2,9 @@
 	import { onMount } from "svelte";
 	import { now } from "svelte/internal";
 
-	import jwt_decode from "jwt-decode";
-
 	import Content from "./Content.svelte";
 	import Message from "./Message.svelte";
+	import SSOButton from "./SOOButton.svelte"
 
 	export let messages: messageType[] = [];
 
@@ -88,53 +87,7 @@
 		//dial();
 		test();
 	});
-
-	function button() {
-		google.accounts.id.renderButton(
-			document.getElementById("googleButton"),
-			{ type: "icon", size: "small" }
-		);
-	}
-
-	async function doPost(data) {
-		let result = null;
-		const res = await fetch("/login", {
-			method: "POST",
-			body: data,
-		});
-
-		const json = await res.json();
-		result = JSON.stringify(json);
-		console.log(result);
-	}
-
-	function handleCredentialResponse(response) {
-		if (response && response.credential) {
-			console.log(response);
-			var decoded = jwt_decode(response.credential);
-			console.log(decoded);
-			doPost(response.credential);
-		}
-	}
-
-	function initSSO() {
-		google.accounts.id.initialize({
-			client_id:
-				"610036027764-0lveoeejd62j594aqab5e24o2o82r8uf.apps.googleusercontent.com",
-			ux_mode: "popup",
-			callback: handleCredentialResponse,
-		});
-		button();
-	}
 </script>
-
-<svelte:head>
-	<script
-		src="https://accounts.google.com/gsi/client"
-		on:load={initSSO}
-		async
-		defer></script>
-</svelte:head>
 
 <header class="site-header">
 	<b>Ochi</b>: find me at
@@ -144,7 +97,7 @@
 	<input bind:value={filter} placeholder="Filter destination port" />
 	<button on:click={filterMessages}>Apply</button>
 	<span>Port number and '&&' to concat.</span>
-	<button id="googleButton">Login with Google</button>
+	<SSOButton />
 </header>
 
 <main>
